**: Clean up copyright notices
A treatise on legal correctness in Copyright notices doesn't belong in a
commit log, but suffice it to say that Linux distributions (Debian in
particular) are insanely pedantic about it as a policy to prevent even
the possibility of a lawsuit.  As a recovering Debianite facing imminent
relapse for Raspple II reasons, I have a strange compulsion to fix such
things.